task3/requirements.txt
pytesseract>=0.3
opencv-python-headless
Pillow
pandas
numpy
tqdm
